1.共识的超时增长时间由2的幂次方改为2的倍数 2.共识的视图编号从byte改为ushort，防止溢出

+++ Zoro/Consensus/ChangeView.cs
@@ -5,7 +5,7 @@ namespace Zoro.Consensus
 {
     internal class ChangeView : ConsensusMessage
     {
-        public byte NewViewNumber;
+        public ushort NewViewNumber;
 
         public ChangeView()
             : base(ConsensusMessageType.ChangeView)
@@ -15,7 +15,7 @@ namespace Zoro.Consensus
         public override void Deserialize(BinaryReader reader)
         {
             base.Deserialize(reader);
-            NewViewNumber = reader.ReadByte();
+            NewViewNumber = reader.ReadUInt16();
             if (NewViewNumber == 0) throw new FormatException();
         }
